backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
google-generativeai
python-dotenv
openai